fix: creating container for Orders, more styling
