+++ src/components/molecules/DragAndDropItem/style.ts
@@ -1,10 +1,4 @@
 const style = () => ({
-    minHeight: '10px',
-    '.hovered': {
-        backgroundColor: 'var(--bs-secondary)',
-        height: '100px',
-        width: 'auto'
-    }
 });
 
 export default style;
