+++ src/components/molecules/DragAndDropZone/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { Row } from 'react-bootstrap';
 import { useDrop } from 'react-dnd'
 import styled, { StyledComponent } from 'styled-components';
 import style from './style'
@@ -7,11 +7,12 @@ interface DragAndDropZone {
     className: string,
     data: any,
     onDrop: Function,
-    itemType: string
+    itemType: string,
+    height?: string
 }
 
 // TODO: make molecule instead of organism, this is reusable
-const DragAndDropZone: StyledComponent = styled(({className, data, onDrop, itemType}) => {
+const DragAndDropZone: StyledComponent = styled(({className, data, onDrop, itemType, height = '10px'}) => {
     const [{ canDrop, isOver }, drop] = useDrop(() => ({
         accept: itemType, // replace with itemType
         collect: (monitor) => ({
@@ -19,7 +20,6 @@ const DragAndDropZone: StyledComponent = styled(({className, data, onDrop, itemT
           canDrop: monitor.canDrop()
         }),
         drop: (item, monitor) => {
-            console.log("dropped, ", item, monitor)
             onDrop(data, item)
         }
     }))
@@ -29,12 +29,9 @@ const DragAndDropZone: StyledComponent = styled(({className, data, onDrop, itemT
     
     // TODO: Items should rearrange inside columns
     return (
-        // TODO, can we have className and hovered both on top level components?
-        <div className={className} ref={drop}>
-            <div
-                className={`'dropZone' ${ isHovered && 'hovered'}`}
-            />
-        </div>
+        <Row className={className} ref={drop} style={{minHeight: height}}>
+            <Row className={`'dropZone' ${ isHovered && 'hovered'}`} />
+        </Row>
     )
 })(style)
