+++ src/components/molecules/KanbanBoardTicketCard/index.tsx
@@ -1,4 +1,4 @@
-import { Button, Card, Col, Container, Row } from 'react-bootstrap';
+import { Card, Row } from 'react-bootstrap';
 import styled, { StyledComponent } from 'styled-components';
 import style from './style';
 import { OrderInterface } from '@interfaces/order';
@@ -10,8 +10,6 @@ interface KanbanBoardTicketCardInterface {
     groupColorMap: any,
     order: OrderInterface
     role: string,
-    // dragRef: any,
-    // color: string
 }
 
 const KanbanBoardTicketCard: StyledComponent = styled(({ 
@@ -19,13 +17,9 @@ const KanbanBoardTicketCard: StyledComponent = styled(({
     groupColorMap,
     order, 
     role, 
-    // dragRef, 
-    // color = '#fff' 
 }: KanbanBoardTicketCardInterface) => {
 
     const [color, setColor] = useState('#fff')
-
-    useEffect(() => console.log(order),[])
 
     const shortenString = (str: string) => {
         if (str.length <= 15) return str
@@ -44,7 +38,7 @@ const KanbanBoardTicketCard: StyledComponent = styled(({
         <Card className={className} role={role}  style={{borderLeft: `${color} 5px solid`}}>
             <Card.Body>
                 <Row>
-                    <p>{order.requester}</p>
+                    <p>Order: {order.id}</p>
                     {/* <p>{order.location.city}</p> */}
                     {/* <Button onClick={() => props.onShowEditTicketCardModal(props.id)}>
                         <EditIcon />
