+++ src/components/organisms/KanbanBoard/index.tsx
@@ -3,10 +3,7 @@ import styled, { StyledComponent } from 'styled-components';
 import style from './style';
 import { OrderInterface } from '@interfaces/order';
 import KanbanBoardTicketCard from '../../molecules/KanbanBoardTicketCard';
-import { useDrag } from 'react-dnd'
 import DragAndDropZone from '@molecules/DragAndDropZone'
-import { colorArray } from '../../../const/colors'
-import { useEffect } from 'react';
 import DragAndDropItem from '@components/molecules/DragAndDropItem';
 
 type BoardType =
@@ -17,7 +14,8 @@ type BoardType =
 interface KanbanBoardInterface {
     className: string,
     groupColorMap: any,
-    orders: { [key: string]: OrderInterface }[]
+    // orders: { [key: string]: OrderInterface }[]
+    orders: any
     type: BoardType,
     setOrderData: React.SetStateAction<any>
 } 
@@ -46,12 +44,11 @@ const KanbanBoard: StyledComponent = styled(({
             return order
         })
         setOrderData(updatedOrders)
-        console.log(updatedOrders)
     }
 
 	return (
 		<Container className={className}>
-            <p>Kanban View - {type}</p>
+            <h3>Kanban View - {type}</h3>
             <Row>
                 {statusOptions.map((status) => (
                     <Col 
@@ -60,28 +57,29 @@ const KanbanBoard: StyledComponent = styled(({
                         >
                         <h5>{status}</h5>
                         {/* TODO: put a ternary here to add a marginTop to KanbanBoardDroopZone where length is 0 */}
-                        <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator' /> 
+                        {/* <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator' />  */}
                             {orders && orders
                                 // @ts-ignore order.status is not read as a string even though it is
                                 .filter((order) => {return order.status === status})
                                 .map((order) => (
-                                        // <DragAndDropItem 
-                                        //     groupColorMap={groupColorMap} 
-                                        //     order={order} type={type} 
-                                        //     key={JSON.stringify(order.id)}/>
-                                    <DragAndDropItem type={type} key={JSON.stringify(order.id)}>
+                                    <Row key={JSON.stringify(order.id)}>
+                                    <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator'/> 
+                                    <DragAndDropItem type={type} key={JSON.stringify(order.id)} orderId={order.id}>
                                         <KanbanBoardTicketCard
                                             groupColorMap={groupColorMap} 
                                             order={order} 
-                                            type={type} 
-                                        
+                                            type={type}
                                         />
                                     </DragAndDropItem>
+                                    </Row>
                                 ))}
                         {/* Only have two KanbanBoardDropZones if orders are in column */}
                         {/* @ts-ignore order.status is not read as a string even though it is */}
-                        {orders && orders.filter((order) => {return order.status === status}).length !==0 &&
-                            <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator' /> }
+                        {/* {console.log(orders, status)} */}
+                        {/* {console.log("test -> ", orders.filter((order: OrderInterface) => {return order.status === status}))} */}
+                        {/* {orders && orders.filter((order: OrderInterface) => {return order.status === status}).length !== 0 &&
+                            <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator' /> } */}
+                        <DragAndDropZone data={{column: status}} onDrop={handleDrop} itemType='Aid Coordinator' height='100%'/>
                     </Col>
                 ))}
             </Row>
