+++ src/components/organisms/OrdersContainer/style.ts
@@ -1,0 +1,14 @@
+const style = () => ({
+    // backgroundColor:'#fff',
+    // margin: '0 0 50px 5em !important',
+    '.orders-controls': {
+        display: 'flex',
+        justifyContent: 'flex-end',
+        marginBottom: '1rem'
+    },
+    '.toggle-view': {
+        marginRight: '1rem'
+    }
+});
+
+export default style;
